integration: move new tests to v2 int tests
